Removed mujoco_env_base.py, added duplicated base env code to the two remaining modules
